+++ R/search_spectra.R
@@ -1,6 +1,20 @@
 #' Search spectra in MSDIAL alignment against database
+#'
+#' This function can be used to identify peaks in your peaktable by matching them
+#' to a spectral database (\code{db}). It takes several arguments that can
+#' be used to customize the matching algorithm, including \code{ri_thresh},
+#' \code{spectral weight}, \code{n_results}. The retention index threshold
+#' (\code{ri_thresh}) is used to subset the provided database, which greatly
+#' improves the search speed. Only database entries with a retention index
+#' falling within the specified threshold will be considered. The spectral
+#' weight affects the weight given to spectral similarity when calculating the
+#' the total similarity score, which is used to rank matches.
+#'
 #' @param x An \code{msdial_alignment} object.
-#' @param db MSP database as list.
+#' @param db MSP database. The provided object should be a nested list, where the
+#' sublists contain the following elements: retention indices in an element named
+#' \code{RI} and mass spectra in an element called \code{Spectra}. All other elements
+#' are optional.
 #' @param cols Index or indices of feature(s) to be identified.
 #' @param ... Additional arguments to \code{\link[OrgMassSpecR]{SpectrumSimilarity}}.
 #' @param ri_thresh Maximum difference between retention indices for a match.
@@ -18,7 +32,11 @@
 #' @note See \href{https://github.com/QizhiSu/mspcompiler}{mspcompiler} for help compiling
 #' an msp database.
 #' @return Returns a modified \code{msdial_alignment} object with database matches
-#' in the \code{matches} slot.
+#' in the \code{matches} slot as a list of data frames. Each \code{data.frame}
+#' will contain the database matches as rows and columns corresponding to the
+#' elements of the database entry (e.g. "Name", "InChIKey", etc.) as well as
+#' match scores for spectral similarity (\code{spectral_match}), retention index
+#' similarity (\code{ri_match}) and the total similarity score (\code{total_score}).
 #' @author Ethan Bass
 #' @export