+++ NEWS.md
@@ -1,4 +1,4 @@
-# msdialreadr 0.3.1 dev
+# msdialreadr 0.3.1
 
 * Made rcdk (for rendering structures in `ms_search_gadget`) suggested instead of
 a dependency.
@@ -7,7 +7,7 @@ types are read in for peak metadata fields. Missing retention indices are now
 correctly entered as NAs.
 * Added error message to `search_spectra` when peak retention indices are not
 present and added option to search entire database by setting `ri_thresh` to `NULL`.
-* Corrected normalizatio functions so they now return the peak table as a
+* Corrected normalization functions so they now return the peak table as a
 data.frame instead of a matrix.
 
 # msdialreadr 0.3.0